Move KinesisShardAssigner to parent package.
